LO-2594: Unit tests, misc bug fixes

+++ limiter/event_processors.py
@@ -2,6 +2,11 @@
 import os
 import sys
 import logging
+from boto3.dynamodb.conditions import Key
+from limiter.clients import dynamodb
+from limiter.managers import RESOURCE_COORDINATE, RESOURCE_ID
+
+logger = logging.getLogger()
 
 class ProcessorPredicate(object):
     def __init__(self, key, pred_func, and_preds=None, or_preds=None):
@@ -19,16 +24,16 @@ class ProcessorPredicate(object):
         return self
 
     def test(self, event):
-        test_value = _reduce_to_path(key)
+        test_value = _reduce_to_path(event, self.key)
 
-        result = self.pred_func(test_value)
+        result = self.pred_func(test_value) if test_value else False
         if self.and_preds and result:
-            for pred in and_preds:
+            for pred in self.and_preds:
                 result &= pred.test(event)
                 if not result:
                     break
-        elif self.or_preds:
-            for pred in and_preds:
+        elif self.or_preds and not result:
+            for pred in self.or_preds:
                 result |= pred.test(event)
                 if result:
                     break
@@ -37,15 +42,17 @@ class ProcessorPredicate(object):
 class EventProcessor(object):
     def __init__(self, source, id_path, predicate=None):
         self.source = source
-        self.predicate = predicate if predicate else lambda x: True
+        self.id_path = id_path
+        self.predicate = predicate
 
     def test_and_get_id(self, event):
-        return _reduce_to_path(id_path) if predicate(event) else None
+        return None if self.predicate and not self.predicate.test(event) else _reduce_to_path(event, self.id_path)
 
 class EventProcessorManager(object):
-    def __init__(self, table_name=None, processors=None):
-        self.processors = {x.source: x, for x in processors} if processors else {}
-        self.table_name = _validate_required_env_fallback(table_name, 'table_name', 'NON_FUNG_TABLE_NAME')
+    def __init__(self, table_name=None, index_name=None, processors=None):
+        self.processors = {x.source: x for x in processors} if processors else {}
+        self.table_name = _validate_required_env_fallback(table_name, 'table_name', 'NON_FUNG_TABLE')
+        self.index_name = _validate_required_env_fallback(index_name, 'index_name', 'NON_FUNG_RES_INDEX')
 
         self._client = None
         self._table = None
@@ -71,22 +78,33 @@ class EventProcessorManager(object):
         processor = self._get_processor(event)
         resource_id = processor.test_and_get_id(event)
         if resource_id:
-            logger.info('Removing {} token {} from {}'.format(processor.source, resource_id, self.table_name))
-            self.table.delete_item(
-                Key={
-                    'resourceId': resource_id
-                }
-            )
+            resource_coordinate = self._get_resource_coordinate(resource_id)
+            if resource_coordinate:
+                logger.info('Removing %s token %s from %s', processor.source, resource_id, self.table_name)
+                self.table.delete_item(
+                    Key={
+                        RESOURCE_COORDINATE: resource_coordinate,
+                        RESOURCE_ID: resource_id
+                    }
+                )
+            else:
+                logger.warn('Could not find a token for resoure %s', resource_id)
 
-    def _get_processor(self, source):
+    def _get_processor(self, event):
         if 'source' not in event:
-            raise ValueError('Cannot process event, source is a required field. Event: ' + event)
+            raise ValueError('Cannot process event, source is a required field. Event: ' + str(event))
 
         source = event['source']
         processor = self.processors.get(source, None)
         if not processor:
             raise ValueError('No processor for event source: ' + source)
         return processor
+
+    def _get_resource_coordinate(self, resource_id):
+        response = self.table.query(
+            IndexName=self.index_name,
+            KeyConditionExpression=Key(RESOURCE_ID).eq(resource_id))
+        return None if response['Count'] == 0 else response['Items'][0][RESOURCE_COORDINATE]
 
 
 # Default processors
@@ -101,10 +119,10 @@ EMR_STEP_COMPLETED = EventProcessor('aws.emr',
 
 BATCH_JOB_COMPLETED = EventProcessor('aws.batch',
                                      'detail.jobId',
-                                      ProcessorPredicate('detail.status',
-                                                         lambda state: state in ['FAILED', 'SUCCEEDED']))
+                                     ProcessorPredicate('detail.status',
+                                                        lambda state: state in ['FAILED', 'SUCCEEDED']))
 
-def _validate_required_env_fallback(self, param_value, param_name, env_var):
+def _validate_required_env_fallback(param_value, param_name, env_var):
     """
     Verify a required argument has a non-null value or has been set via an environment variable.
